windows-x86: skip the check on compiled dependencies with --ignore-build=all
Rationale: To prepare the windows build, both the dependencies and nidium's code needs to be ported.
This is a lot of work. With this change it is possible to split this up:
The command 'configure_libnidiumcore --ignore-build=all' will:
* download the dependencies, but will not build them.
* create the gyp files
* create the visual studio files
* visual studio can compile but WILL fail on linking

This means that the visual studio tools can be used to port nidiums cpp code.

The visual studio solution (and also project files) are located in .\gyp\*.sln