windows-x86: let the konstructor script determine the platform type
Problem: Gyp builds the msvc project files based on these settings.

With this change konstructor can determine and set the platform also for 32 bit systems.
Please note: although this is, slightly better then it was before, it is not ideal for other processors.
(arm, mips, ..). This might also effect crosscompiling (android?).